Add gpgv --keyring
gpg installed keys in pubring.kbx and gpgv tried to find them in
the trustedkeys.kbx. Signatures did not verify.

In previous commit, the KEYS were updated from
https://downloads.apache.org/qpid/KEYS.

The excuse for gpgv was wrong. gpg complained of keys not being trusted,
but exit code was 0. gpgv treats all the keys in trustedkeys.kbx
trustworthy and does not complain.